requirements.txt
tqdm==4.44.1
opencv-python==4.2.0.34
# Shapely==1.6.4.post1
matplotlib==3.1.3
numpy==1.18.3
tensorflow==2.0.0
Cython==0.29.15
# ipython==5.1.0
pytesseract==0.3.4   
pillow==7.0.0 
flask==1.1.2 
flask-cors==3.0.8